refactor: 도메인의 repository 패키지 변경
