+++ kitchenpos-core/src/main/java/kitchenpos/menugroup/repository/MenuGroupRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.menugroup.domain.repository;
+package kitchenpos.menugroup.repository;
 
 import kitchenpos.menugroup.domain.MenuGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
