+++ kitchenpos-core/src/main/java/kitchenpos/order/application/OrderService.java
@@ -2,18 +2,18 @@ package kitchenpos.order.application;
 
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.order.domain.MenuSnapshot;
-import kitchenpos.menu.domain.repository.MenuRepository;
+import kitchenpos.menu.repository.MenuRepository;
 import kitchenpos.menu.exception.NotExistMenuException;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderLineItem;
 import kitchenpos.order.domain.OrderStatus;
-import kitchenpos.order.domain.repository.OrderRepository;
+import kitchenpos.order.repository.OrderRepository;
 import kitchenpos.order.dto.OrderChangeStatusRequest;
 import kitchenpos.order.dto.OrderCreateRequest;
 import kitchenpos.order.dto.OrderLineItemRequest;
 import kitchenpos.order.exception.NotExistOrderException;
 import kitchenpos.ordertable.domain.OrderTable;
-import kitchenpos.ordertable.domain.repository.OrderTableRepository;
+import kitchenpos.ordertable.repository.OrderTableRepository;
 import kitchenpos.ordertable.exception.NotExistOrderTable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
