+++ kitchenpos-core/src/main/java/kitchenpos/order/repository/OrderLineItemRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.domain.repository;
+package kitchenpos.order.repository;
 
 import kitchenpos.order.domain.OrderLineItem;
 import org.springframework.data.jpa.repository.JpaRepository;
