+++ kitchenpos-core/src/test/java/kitchenpos/application/MenuServiceTest.java
@@ -10,10 +10,10 @@ import kitchenpos.menu.exception.InvalidMenuPriceException;
 import kitchenpos.product.exception.InvalidProductPriceException;
 import kitchenpos.menugroup.exception.NotExistMenuGroupException;
 import kitchenpos.menu.application.MenuService;
-import kitchenpos.menugroup.domain.repository.MenuGroupRepository;
-import kitchenpos.menu.domain.repository.MenuProductRepository;
-import kitchenpos.menu.domain.repository.MenuRepository;
-import kitchenpos.product.domain.repository.ProductRepository;
+import kitchenpos.menugroup.repository.MenuGroupRepository;
+import kitchenpos.menu.repository.MenuProductRepository;
+import kitchenpos.menu.repository.MenuRepository;
+import kitchenpos.product.repository.ProductRepository;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
