+++ kitchenpos-core/src/test/java/kitchenpos/application/OrderTableServiceTest.java
@@ -7,17 +7,17 @@ import kitchenpos.ordertable.dto.OrderTableCreateRequest;
 import kitchenpos.ordertable.exception.InvalidOrderTableChangeEmptyException;
 import kitchenpos.ordertable.exception.InvalidOrderTableChangeNumberOfGuestsException;
 import kitchenpos.menu.domain.Menu;
-import kitchenpos.menu.domain.repository.MenuRepository;
+import kitchenpos.menu.repository.MenuRepository;
 import kitchenpos.menugroup.domain.MenuGroup;
-import kitchenpos.menugroup.domain.repository.MenuGroupRepository;
+import kitchenpos.menugroup.repository.MenuGroupRepository;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderLineItem;
-import kitchenpos.order.domain.repository.OrderRepository;
+import kitchenpos.order.repository.OrderRepository;
 import kitchenpos.ordertable.application.OrderTableService;
 import kitchenpos.ordertable.domain.OrderTable;
-import kitchenpos.ordertable.domain.repository.OrderTableRepository;
+import kitchenpos.ordertable.repository.OrderTableRepository;
 import kitchenpos.tablegroup.domain.TableGroup;
-import kitchenpos.tablegroup.domain.repository.TableGroupRepository;
+import kitchenpos.tablegroup.repository.TableGroupRepository;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
